+++ lib/formatters/credit_card_expiration_input_formatter.dart
@@ -23,19 +23,10 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-import 'dart:math';
-
-import 'package:flutter/services.dart';
 
 import 'formatter_utils.dart';
 import 'masked_input_formatter.dart';
 
-class CreditCardNumberFormatter extends MaskedInputFormater {
-  CreditCardNumberFormatter() : super('0000 0000 0000 0000');
-}
-class CvvCodeFormatter extends MaskedInputFormater {
-  CvvCodeFormatter() : super('000');
-}
 
 class CreditCardExpirationDateFormatter extends MaskedInputFormater {
   CreditCardExpirationDateFormatter() : super('00/00');
@@ -77,37 +68,4 @@ class CreditCardExpirationDateFormatter extends MaskedInputFormater {
     }
     return result;
   }
-}
-
-/// allows only latin characters and converts them to uppercase
-/// you can use TextCapitalization.characters instead of this formatter 
-/// in most cases
-class CreditCardHolderNameFormatter extends TextInputFormatter {
-
-  static RegExp _nameMatcher = RegExp(r'[A-Z ]+');
-
-  @override
-  TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
-    var isErasing = newValue.text.length < oldValue.text.length;
-    if (isErasing) {
-      return newValue;
-    } 
-    var newText = newValue.text.toUpperCase();
-    var text = newText.split('')
-      .where((s) => _nameMatcher.hasMatch(s))
-      .map((s) => s.toUpperCase())
-      .join('');
-    var endOffset = max(oldValue.text.length - oldValue.selection.end, 0);
-    var selectionEnd = text.length - endOffset;
-    
-    return TextEditingValue(
-      composing: TextRange.collapsed(selectionEnd),
-      selection: TextSelection.collapsed(
-        offset: selectionEnd, 
-        affinity: TextAffinity.downstream
-      ),
-      text: text,
-    );
-  }
-  
 }